+++ gufe/mapping/__init__.py
@@ -3,7 +3,7 @@
 """Defining the relationship between different components"""
 
 from .atom_mapper import AtomMapper
-from .atom_mapping import AtomMapping, LigandAtomMapping
+from .atom_mapping import AtomMapping
 from .componentmapping import ComponentMapping
 from .errors import AtomMappingError
 from .ligandatommapping import LigandAtomMapping
